first attempt (See commit)
So I tried to package the first kr package, but it seems to be needing
the package sigchain as evidence by the Makefile[0]. I then tried to
package sigchain but I quickly realized, given the error message below,
that packages relies on each other within the same source tree. Can I
set the source of one package to be the modified source of the previous
package? Essentially I do not really care about the output of the
previous derivation and should probably set the installPhase to `true`.

[0]: https://github.com/kryptco/kr/blob/ee6b7db8abd009f0dd89d7175ee27c09ff74fed4/Makefile#L57-L58