build-bazel-package: replace the sha256 in all marker files by 1
This commit forces the marker files to be deterministic. It seems that
Bazel computes the sha256 of all dependencies without removing the .git
folders which would result in the derivation getting broken as soon as
the dependencies are updated upstream. Fortunately, Bazel does not
validate the sha256 during the build phase.